Extract functions into modules in prep for transform plugin
